LayoutTests/fast/inspector-support/uncaught-dom1-exception-expected.txt

CONSOLE MESSAGE: line 12: IndexSizeError (DOM Exception 1): The index is not in the allowed range.
This tests that we are getting the correct message for DOM Exception 1: INDEX_SIZE_ERR.
